+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Display group config in a group header (roubkar, gorarakelyan)
 - Sort runs by date in Dashboard screen (roubkar)
 - Add offset for grouped items (roubkar)
+- Add group context info to context table (roubkar)
 
 ## 1.1.10 Nov 5 2020
 - Expand group and scroll to the corresponding item, when clicking on a line chart (roubkar)